requirements.txt
PasteScript==3.2.1
plaster-pastedeploy==0.7
pyramid==1.10.5
pyramid-chameleon==0.3
pyramid-debugtoolbar==4.8
waitress==1.4.4
python-gammu==3.1
colander==1.8.3
phonenumbers==8.12.24
deform==2.0.15
